Refatoração + divisão em arquivos + rota para consultar um registro do kv
