+++ src/constants.js
@@ -1,0 +1,5 @@
+export const statusCode = 302;
+
+export const CLICKS_NAMESPACE = 'clicks:';
+
+export const invalidPaths = ['shorten', 'api', 'login', 'del', 'edit', 'links', 'admin', 'dashboard', 'settings'];
