+++ src/utils/format-expiration-time.js
@@ -1,0 +1,6 @@
+export function formatExpirationTime(expirationTime) {
+	const now = new Date();
+	const expirationDate = new Date(now.getTime() + expirationTime * 60 * 1000);
+  
+	return expirationDate.getTime();
+}
