+++ src/utils/generate-random-key.js
@@ -1,0 +1,11 @@
+export function generateRandomKey(length) {
+	const characters = '23456789abcdefghijkmnopqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ';
+
+	let result = '';
+
+	for(let i = 0; i < length; i++) {
+		result += characters.charAt(Math.floor(Math.random() * characters.length));
+	}
+  
+	return result;
+}
