+++ src/utils/generate-unique-key.js
@@ -1,0 +1,6 @@
+export function generateUniqueKey() {
+	const timestamp = Date.now();
+	const randomString = Math.random().toString(36).substring(2, 12);
+  
+	return `${timestamp}${randomString}`;
+}
